Renaming contacts folders to contacts-management
