+++ spec/components/schemas/contacts-management/contact-list.ts
