+++ spec/paths/contacts/contact-lists.ts
@@ -1,6 +1,6 @@
 import { PathItemObject, OperationObject, ResponseObject, ResponsesObject } from 'openapi3-ts';
 import { ref as errorResponseRef } from '../../components/responses/error';
-import { ref as contactListRef } from '../../components/schemas/contacts/contact-list';
+import { ref as contactListRef } from '../../components/schemas/contacts-management/contact-list';
 import { ref as pageRef } from '../../components/parameters/page';
 import { ref as sizeRef } from '../../components/parameters/size';
